Migrate dotted_border and dots_indicator to their new APIs
Both broke in the major-version upgrade.

dotted_border 3.x moved styling out of the widget constructor into an
options object, so dashPattern, borderType, radius and color are now
fields of RoundedRectDottedBorderOptions, which supplies BorderType.RRect
itself. The widget's new ignoring flag defaults to true but wraps only
the border painter, and childOnTop defaults to true, so the GestureDetector
inside each of these two borders still receives taps.

dots_indicator 5.x narrowed position from num to double, so the page
index needs an explicit toDouble().

flutter analyze: 0 errors. flutter test: 38 passing.

+++ lib/ui/screens/intro/intro_screen.dart
@@ -79,7 +79,7 @@ class _IntroScreenState extends State<IntroScreen> {
                     builder: (context, snapshot) {
                       return DotsIndicator(
                         dotsCount: 3,
-                        position: snapshot.data ?? 0,
+                        position: (snapshot.data ?? 0).toDouble(),
                         decorator: DotsDecorator(
                           color: const Color(0xffEEE5FF),
                           activeColor: violet100,
